# decant-standalone/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Decant API
  description: |
    RESTful API for managing hierarchical knowledge graphs with AI-powered classification and enrichment.

    Decant provides dual-hierarchy organization (functional and organizational), automatic content
    classification using GPT-4o-mini, and real-time updates via Server-Sent Events.

    ## Key Features
    - Dual hierarchy organization (function & organization)
    - AI-powered content classification
    - Phase 2 enrichment for deep analysis
    - Full-text search with advanced filters
    - Real-time notifications via SSE
    - Backup and restore capabilities
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: http://localhost:{port}
    description: Configurable port
    variables:
      port:
        default: '3000'
        description: Server port
...
paths:
...
  /api/nodes:
    get:
      tags: [Nodes]
      summary: Get all nodes
      description: Retrieve all nodes with optional pagination
      operationId: getAllNodes
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Node'
                  - $ref: '#/components/schemas/PaginatedNodesResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Nodes]
      summary: Create a new node
      description: Create a new node in the knowledge graph
      operationId: createNode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNodeInput'
      responses:
        '201':
          description: Node created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/nodes/{id}:
    get:
      tags: [Nodes]
      summary: Get a single node
      description: Retrieve a specific node by ID
      operationId: getNode
      parameters:
        - $ref: '#/components/parameters/NodeIdParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [Nodes]
      summary: Update a node
      description: Update an existing node's properties
      operationId: updateNode
      parameters:
        - $ref: '#/components/parameters/NodeIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNodeInput'
      responses:
        '200':
          description: Node updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Nodes]
      summary: Delete a node
      description: Soft-delete a node (marks as deleted)
      operationId: deleteNode
      parameters:
        - $ref: '#/components/parameters/NodeIdParam'
      responses:
        '200':
          description: Node deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/nodes/{id}/merge:
    post:
      tags: [Nodes]
      summary: Merge two nodes
      description: Merge a secondary node into the primary node
      operationId: mergeNodes
      parameters:
        - $ref: '#/components/parameters/NodeIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MergeNodesInput'
      responses:
        '200':
          description: Nodes merged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/nodes/{id}/move:
    post:
      tags: [Nodes]
      summary: Move a node
      description: Move a node to a new parent in the hierarchy
      operationId: moveNode
      parameters:
        - $ref: '#/components/parameters/NodeIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveNodeInput'
      responses:
        '200':
          description: Node moved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/nodes/{id}/related:
    get:
      tags: [Nodes]
      summary: Get related nodes
      description: Get nodes similar to the specified node
      operationId: getRelatedNodes
      parameters:
        - $ref: '#/components/parameters/NodeIdParam'
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 20
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelatedNodesResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/search/advanced:
    get:
      tags: [Search]
      summary: Advanced search
      description: Advanced search with filters and faceted results
      operationId: searchAdvanced
      parameters:
        - name: q
          in: query
          required: true
          description: Search query string
          schema:
            type: string
        - name: segment
          in: query
          description: Filter by segment code
          schema:
            type: string
        - name: category
          in: query
          description: Filter by category code
          schema:
            type: string
        - name: contentType
          in: query
          description: Filter by content type
          schema:
            type: string
        - name: organization
          in: query
          description: Filter by organization (partial match)
          schema:
            type: string
        - name: dateStart
          in: query
          description: Filter by start date (ISO format)
          schema:
            type: string
            format: date-time
        - name: dateEnd
          in: query
          description: Filter by end date (ISO format)
          schema:
            type: string
            format: date-time
        - name: hasMetadata
          in: query
          description: Only enriched nodes
          schema:
            type: boolean
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvancedSearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================
  # Import API
  # ============================================================
...
components:
  parameters:
    NodeIdParam:
      name: id
      in: path
      required: true
      description: Node UUID
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 500
  schemas:
...
    Node:
      type: object
      required:
        - id
        - title
        - url
        - source_domain
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        url:
          type: string
          format: uri
        source_domain:
          type: string
        company:
          type: string
          nullable: true
        phrase_description:
          type: string
          nullable: true
        short_description:
          type: string
          nullable: true
        logo_url:
          type: string
          format: uri
          nullable: true
        ai_summary:
          type: string
          nullable: true
        extracted_fields:
          type: object
          additionalProperties: true
        metadata_tags:
          type: array
          items:
            type: string
        key_concepts:
          type: array
          items:
            type: string
        function_parent_id:
          type: string
          format: uuid
          nullable: true
        organization_parent_id:
          type: string
          format: uuid
          nullable: true
        function_code:
          type: string
          nullable: true
        organization_code:
          type: string
          nullable: true
        phase2_completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateNodeInput:
      type: object
      required:
        - title
        - url
        - source_domain
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        url:
          type: string
          format: uri
        source_domain:
          type: string
          minLength: 1
        company:
          type: string
          maxLength: 200
        phrase_description:
          type: string
          maxLength: 1000
        short_description:
          type: string
          maxLength: 2000
        logo_url:
          type: string
          format: uri
          nullable: true
        ai_summary:
          type: string
          maxLength: 5000
        extracted_fields:
          type: object
        metadata_tags:
          type: array
          items:
            type: string
        key_concepts:
          type: array
          items:
            type: string
        function_parent_id:
          type: string
          format: uuid
          nullable: true
        organization_parent_id:
          type: string
          format: uuid
          nullable: true
    UpdateNodeInput:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        company:
          type: string
          maxLength: 200
        phrase_description:
          type: string
          maxLength: 1000
        short_description:
          type: string
          maxLength: 2000
        logo_url:
          type: string
          format: uri
          nullable: true
        ai_summary:
          type: string
          maxLength: 5000
        extracted_fields:
          type: object
        metadata_tags:
          type: array
          items:
            type: string
        key_concepts:
          type: array
          items:
            type: string
        function_parent_id:
          type: string
          format: uuid
          nullable: true
        organization_parent_id:
          type: string
          format: uuid
          nullable: true
    MergeNodesInput:
      type: object
      required:
        - secondaryId
      properties:
        secondaryId:
          type: string
          format: uuid
        options:
          type: object
          properties:
            keepMetadata:
              type: boolean
            appendSummary:
              type: boolean
    MoveNodeInput:
      type: object
      required:
        - targetHierarchy
      properties:
        targetParentId:
          type: string
          format: uuid
          nullable: true
        targetHierarchy:
          type: string
          enum: [function, organization]
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: string
    PaginatedNodesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Node'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    RelatedNodesResponse:
      type: object
      properties:
        nodeId:
          type: string
          format: uuid
        related:
          type: array
          items:
            type: object
            properties:
              node:
                type: object
                properties:
                  id:
                    type: string
                  title:
                    type: string
                  url:
                    type: string
                  segment:
                    type: string
                  category:
                    type: string
                  contentType:
                    type: string
                  logo_url:
                    type: string
                  phrase_description:
                    type: string
              similarityScore:
                type: integer
                minimum: 0
                maximum: 100
              sharedAttributes:
                type: array
                items:
                  type: string
    SearchResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        url:
          type: string
        company:
          type: string
        phrase_description:
          type: string
        extracted_fields:
          type: object
        matchScore:
          type: number
          format: float
...
    AdvancedSearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        facets:
          type: object
          properties:
            segments:
              type: object
              additionalProperties:
                type: integer
            categories:
              type: object
              additionalProperties:
                type: integer
            contentTypes:
              type: object
              additionalProperties:
                type: integer
            organizations:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  count:
                    type: integer
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication (future enhancement)